% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bash_ref.R
\name{build_bash_ref}
\alias{build_bash_ref}
\title{Build Reference Page For Bash Script Documentations}
\usage{
build_bash_ref(ps_script_path, ps_out_dir)
}
\arguments{
\item{ps_script_path}{}

\item{ps_out_dir}{}
}
\description{
Bash scripts given in a vector to ps_script_path are converted into
html-pages. These pages are moved to an output directory. An index
page is constructed which can be referred by other pages.
}
\examples{
\dontrun{
build_bash_ref(ps_script_path = file.path('inst', 'bash'), 
               ps_out_dir     = file.path('docs', 'bash_ref') )
}

}
