% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bash_ref.R
\name{get_title}
\alias{get_title}
\title{Get Title From HTML File}
\usage{
get_title(ps_html_path)
}
\arguments{
\item{ps_html_path}{}
}
\value{
s_cur_title Title from html page
}
\description{
The html page generated from a script contains a title between html-tags.
This function searches for the title-tag and returns the string that is
between the tags.
}
\examples{
\dontrun{
get_title(ps_html_path = 'docs/bash_ref/new_bash_script.html')
}

}
