% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_navbar.R
\name{template_navbar_bash_ref}
\alias{template_navbar_bash_ref}
\title{Create Navbar YAML Template}
\usage{
template_navbar_bash_ref(
  ps_pkgdown_path = ".",
  ps_nb_yml_outfile = "_navbar.yml",
  pb_force_output = FALSE
)
}
\arguments{
\item{ps_pkgdown_path}{path of pkgdown home}

\item{ps_nb_yml_outfile}{navbar output file}

\item{pb_force_output}{overwrite existing output file}
}
\description{
For a given project containing bash scripts for which documentations were
generated, a template for a yaml-definition-file for the common navbar is
created. The required information is taken from the output of pkgdown::as_pkgdown.
}
\examples{
\dontrun{
template_navbar_bash_ref(ps_pkgdown_path = '.',
                         ps_nb_yml_outfile = 'pkgdown/_navbar.yml')
}

}
